김광민 마이페이지, 관리자페이지 인터페이스 변경
