+++ BearBom/src/main/frontend/src/PageComponents/MyPage/Menu/WishlistView.jsx
@@ -1,0 +1,14 @@
+import React, { useState } from "react";
+import Table from "../../../ModuleComponents/Table/Table";
+
+const WishlistView = () => {
+  const [tableItems, setTableItems] = useState();
+
+  return (
+    <>
+      <Table tableItems={tableItems} />
+    </>
+  );
+};
+
+export default WishlistView;
